// src/app/models/query-models/step.ts
/**
 * Copyright 2017 - 2018  The Hyve B.V.
 *
 * This Source Code Form is subject to the terms of the Mozilla Public
 * License, v. 2.0. If a copy of the MPL was not distributed with this
 * file, You can obtain one at http://mozilla.org/MPL/2.0/.
 */

/**
 * A helper class to denote the step
 * where the user is currently composing her query
 */
export enum Step {
  I = 'I',
  II = 'II'
}