// src/app/models/query-models/query-subscription-frequency.ts
/**
 * Copyright 2017 - 2018  The Hyve B.V.
 *
 * This Source Code Form is subject to the terms of the Mozilla Public
 * License, v. 2.0. If a copy of the MPL was not distributed with this
 * file, You can obtain one at http://mozilla.org/MPL/2.0/.
 */

export enum QuerySubscriptionFrequency {
  DAILY = 'DAILY',
  WEEKLY = 'WEEKLY'
}